Title: 특정 문자로 시작하는 문자열, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 250918/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.cpp
@@ -12,7 +12,7 @@ int main() {
         cin >> str[i];
     char c;
     cin >> c;
-    
+
     int cnt = 0;
     int sum = 0;
     for(int i=0;i<n;i++)
@@ -23,6 +23,7 @@ int main() {
             sum += str[i].length();
         }
     }
-    cout << cnt << " " << fixed << setprecision(2) << double(sum/cnt);
+    
+    cout << cnt << " " << fixed << setprecision(2) << double(sum)/double(cnt);
     return 0;
 }
